Brought lesson up to date with current style
I merged `lesson-template`, which brought in all of the new CSS
and other style changes. This merge, and the fact that these files
haven't been looked at in a while, highlighted some things that this
repo is missing. I attempted to fix all of the issues that were not
too time consuming. With this, the only errors / warnings from
`tools/check.py` are issues with structure, or missing content.
The generated HTML files should look much more consistent with other
lessons, as well.

Specifically, I:
- Added time estimates to all lessons (60 minutes each)
- Fixed styling of challenges, lesson objectives, etc.
- Separated challenges into separate callouts with titles
- Fixed code cells that used the wrong CSS class, or didn't use
  the tilde-fencing properly
- Removed excessive whitespace
- Fixed external links in multiprocessing
- Fixed links to wrong example location in multiprocessing